Added flexible login information setting
Moved the login credentials into settings where they can be easily found
and adjusted. Added ability to specify live/demo account. without this
"live" was hard-coded into code and made it impossible to use
application with a demo account.

Also added exception handling in order to actually display the login
error to the user. Before this it failed silently, very confusing and
time consuming to debug.